ReactiveMongoTemplate 사용하여 DatabaseInitializer를 CustomerRepository로 대체
